pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "symphony"
version = "0.2.0"
description = "Multi-project Linear orchestrator — poll tracker, dispatch coding agents"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "structlog>=24.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "watchfiles>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.scripts]
symphony = "symphony.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/symphony"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
